app/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
pymongo==3.11.4
PyMySQL==1.0.2
SQLAlchemy==1.4.15
Werkzeug==2.0.1
